Add: PayloadController追加、Payload作成部分を移行
